+++ xwalk_android.gypi
@@ -180,7 +180,7 @@
         'runtime/android/core_internal/src/org/xwalk/core/internal/XWalkContentsIoThreadClient.java',
         'runtime/android/core_internal/src/org/xwalk/core/internal/XWalkCookieManagerInternal.java',
         'runtime/android/core_internal/src/org/xwalk/core/internal/XWalkDevToolsServer.java',
-        'runtime/android/core_internal/src/org/xwalk/core/internal/XWalkHttpAuthHandler.java',
+        'runtime/android/core_internal/src/org/xwalk/core/internal/XWalkHttpAuthHandlerInternal.java',
         'runtime/android/core_internal/src/org/xwalk/core/internal/XWalkPathHelper.java',
         'runtime/android/core_internal/src/org/xwalk/core/internal/XWalkSettings.java',
         'runtime/android/core_internal/src/org/xwalk/core/internal/XWalkViewDelegate.java',
